refactor state organization
