+++ cosmos.config.json
@@ -1,6 +1,7 @@
 {
   "watchDirs": ["src/lib"],
   "webpack": {
+    "port": 5001,
     "configPath": "react-scripts/config/webpack.config",
     "overridePath": "cosmos.override.js"
   }
