+++ src/lib/components/Swap/Settings/TransactionTtlInput.tsx
@@ -1,12 +1,12 @@
 import { Trans } from '@lingui/macro'
 import { useAtom } from 'jotai'
+import { TRANSACTION_TTL_DEFAULT, transactionTtlAtom } from 'lib/state/swap'
 import styled, { ThemedText } from 'lib/theme'
 import { useRef } from 'react'
 
 import Column from '../../Column'
 import { inputCss, IntegerInput } from '../../Input'
 import Row from '../../Row'
-import { TRANSACTION_TTL_DEFAULT, transactionTtlAtom } from '../state'
 import { Label } from './components'
 
 const tooltip = <Trans>Your transaction will revert if it has been pending for longer than this period of time.</Trans>
