+++ src/lib/components/Swap/Settings/index.tsx
@@ -2,6 +2,7 @@ import { Trans } from '@lingui/macro'
 import { useResetAtom } from 'jotai/utils'
 import useScrollbar from 'lib/hooks/useScrollbar'
 import { Settings as SettingsIcon } from 'lib/icons'
+import { settingsAtom } from 'lib/state/swap'
 import styled, { ThemedText } from 'lib/theme'
 import React, { useState } from 'react'
 
@@ -9,7 +10,6 @@ import { IconButton, TextButton } from '../../Button'
 import Column from '../../Column'
 import Dialog, { Header } from '../../Dialog'
 import { BoundaryProvider } from '../../Popover'
-import { settingsAtom } from '../state'
 import MaxSlippageSelect from './MaxSlippageSelect'
 import TransactionTtlInput from './TransactionTtlInput'
