+++ src/lib/components/Swap/Summary/index.tsx
@@ -3,6 +3,7 @@ import { useAtomValue } from 'jotai/utils'
 import { IconButton } from 'lib/components/Button'
 import useScrollbar from 'lib/hooks/useScrollbar'
 import { Expando, Info } from 'lib/icons'
+import { Input, inputAtom, outputAtom, swapAtom } from 'lib/state/swap'
 import styled, { ThemedText } from 'lib/theme'
 import { useMemo, useState } from 'react'
 
@@ -11,7 +12,6 @@ import Column from '../../Column'
 import { Header } from '../../Dialog'
 import Row from '../../Row'
 import Rule from '../../Rule'
-import { Input, inputAtom, outputAtom, swapAtom } from '../state'
 import Details from './Details'
 import Summary from './Summary'
